% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcoint_fcoint.R
\name{print.fcoint}
\alias{print.fcoint}
\title{Print Method for fcoint Objects}
\usage{
\method{print}{fcoint}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"fcoint"}.}

\item{...}{Further arguments passed to or from other methods (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for fcoint Objects
}
