% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtcadfcoint_xtcadfcoint.R
\name{print.xtcadfcoint}
\alias{print.xtcadfcoint}
\title{Print method for xtcadfcoint objects}
\usage{
\method{print}{xtcadfcoint}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"xtcadfcoint"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for xtcadfcoint objects
}
