% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtcadfcoint_xtcadfcoint.R
\name{summary.xtcadfcoint}
\alias{summary.xtcadfcoint}
\title{Summary method for xtcadfcoint objects}
\usage{
\method{summary}{xtcadfcoint}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"xtcadfcoint"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{object}.
}
\description{
Summary method for xtcadfcoint objects
}
